Update flax_stable_noClock.lua
Added checkbox for safeClick - safeClick() part of common files is designed to help prevent misclicks when players accidentally move mouse while running.

We think that when using on Multi-Monitors that this sometimes causes an occasional misclick when dragging windows.  This is a test feature to see if this can help resolve this issue.